Add files via upload
The shape and size of template are very important for data event extraction. Inspired by the suggestion of a Anonymous  reviewer, we added two parameters which are template type and template file. Users can define a set of simple templates by giving the number of cell in x, y, and z direction (template type=0) or define complex templates with the needed shape by giving the relative coordinates of template elements in a data template file.